requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.7
# by the following command:
#
#    pip-compile
#
build==0.10.0
    # via pip-tools
cfgv==3.3.1
    # via pre-commit
click==8.1.3
    # via pip-tools
distlib==0.3.6
    # via virtualenv
filelock==3.9.0
    # via virtualenv
identify==2.5.13
    # via pre-commit
importlib-metadata==6.0.0
    # via
    #   build
    #   click
    #   pre-commit
    #   virtualenv
jinja2==3.1.2
    # via -r requirements.in
markupsafe==2.1.2
    # via jinja2
mitogen==0.3.3
    # via -r requirements.in
nodeenv==1.7.0
    # via pre-commit
packaging==23.0
    # via build
pip-tools==6.12.1
    # via -r requirements.in
platformdirs==2.6.2
    # via virtualenv
pre-commit==2.21.0
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via build
pyyaml==6.0
    # via
    #   -r requirements.in
    #   pre-commit
tomli==2.0.1
    # via
    #   build
    #   pyproject-hooks
typing-extensions==4.4.0
    # via
    #   importlib-metadata
    #   platformdirs
virtualenv==20.17.1
    # via pre-commit
wheel==0.38.4
    # via pip-tools
zipp==3.11.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
